rocket-rag-backend/requirements.txt
fastapi
uvicorn
langchain
chromadb
python-dotenv
openai
huggingface-hub
sentence-transformers
